refactor: change latency type in the heartbeat table

+++ src/main/java/io/uptimego/model/Heartbeat.java
@@ -28,7 +28,7 @@ public class Heartbeat {
     @Enumerated(EnumType.STRING)
     private HeartbeatStatus status;
 
-    private double latency;
+    private Integer latency;
 
     @CreationTimestamp
     private LocalDateTime timestamp;
